requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
cloudinary==1.29.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.0.6
django-ckeditor==6.4.2
django-js-asset==2.0.0
django-taggit==3.0.0
gunicorn==20.1.0
idna==3.3
Pillow==9.2.0
psycopg2==2.9.3
python-decouple==3.6
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
static3==0.7.0
tzdata==2022.1
urllib3==1.26.10
